+++ force-app/main/default/classes/RestaurantsController.cls
@@ -1,12 +1,12 @@
 public with sharing class RestaurantsController {
     @AuraEnabled
-    public static List<Restaurant__c> getRestaurants(ID tripId) {
+    public static List<Restaurant__c> getRestaurants(Id tripId) {
         List<Restaurant__c> restaurants = new List<Restaurant__c>();
         Restaurant__c restaurant = new Restaurant__c();
         Location randomLoc = Location.newInstance(28.635308,77.22496);
         restaurant.Location__Latitude__s = randomLoc.latitude;
         restaurant.Location__Longitude__s = randomLoc.longitude;
-        restaurant.Name = 'Test';
+        restaurant.Name = 'Tes123t';
         restaurant.Travel_Plan__c = tripId;
         restaurants.add(restaurant);
         System.debug(restaurant.Location__Latitude__s);
