+++ force-app/main/default/classes/TravelAppTripController.cls
@@ -1,0 +1,2 @@
+public with sharing class TravelAppTripController {
+}
