--- force-app/main/default/lwc/travelApp/travel_app.html
@@ -1,20 +1,0 @@
-<template>
-    <lightning-card title="Travel Itinerary">
-        <lightning-button
-            onclick={handleClick}
-            aria-haspopup="dialog"
-            label="Create New Trip"
-            slot="actions">
-        </lightning-button>
-        <!-- <lightning-button label="New Trip" slot="actions" onclick={handleClick}></lightning-button> -->
-        <!-- Later put this as a list that pulls from a database of current active trips -->
-        <lightning-tabset>
-            <lightning-tab label="Japan">
-                <c-travel-app-trip></c-travel-app-trip>
-            </lightning-tab>
-            <lightning-tab label="China">
-                <c-travel-app-trip></c-travel-app-trip>
-            </lightning-tab>
-        </lightning-tabset>
-    </lightning-card>
-</template>
