--- force-app/main/default/lwc/travel_app/travel_app.js-meta.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>56.0</apiVersion>
-    <isExposed>true</isExposed>
-    <targets>
-        <target>lightning__RecordPage</target>
-        <target>lightning__HomePage</target>
-        <target>lightning__Tab</target>
-        <!-- Add more targets as needed -->
-    </targets>
-</LightningComponentBundle>
